fix(entity): Change Lead.stage from Integer to Short for SMALLINT compatibility
Hibernate schema validation requires Short type to match PostgreSQL SMALLINT.
Migration V255 defines stage as SMALLINT, not INTEGER.

Changes:
- Lead.stage: Integer → Short
- LeadProtectionService methods work with int (auto-conversion)
- No test changes needed (Pure Mockito tests)

Fixes: Schema-validation error on backend startup
Refs: Sprint 2.1.5, V255, ADR-004

+++ backend/src/main/java/de/freshplan/modules/leads/domain/Lead.java
@@ -179,7 +179,7 @@ public class Lead extends PanacheEntityBase {
 
   // Progressive Profiling Stage (Sprint 2.1.5 - V255)
   @Column(name = "stage", nullable = false)
-  public Integer stage = 0; // 0=Vormerkung, 1=Registrierung, 2=Qualifiziert
+  public Short stage = 0; // 0=Vormerkung, 1=Registrierung, 2=Qualifiziert
 
   // Metadata
   @Size(max = 100)
